Ak feature

+++ src/main/java/pro/sky/telegrambotshelter/service/ShelterService.java
@@ -13,7 +13,7 @@ public class ShelterService {
     }
 
     /**метод для получения id волонтера*/
-    public int getVolunteerChatId(String shelterType) {
+    public long getVolunteerChatId(String shelterType) {
         return shelterRepository.findShelterByShelterType(shelterType).getVolunteerChatId();
     }
